// src/types/api.ts
export interface components {
    schemas: {
// ...
        SpeechTimestamp: {
            /** Start */
            start: number;
            /** End */
            end: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}